network battery dummy table add

+++ db/migrate/20131025120940_create_components.rb
@@ -2,7 +2,7 @@ class CreateComponents < ActiveRecord::Migration
   def change
     create_table :components do |t|
       t.string :name
-      t.decimal :joule
+      t.float :joule
       t.references :battery
 
       t.timestamps
